Se crea el hook useProjectsData para separar parte de la logica de select project
