+++ FrontEnd/src/Pages/payrollProjects/SelectProject.js
@@ -1,63 +1,24 @@
 import { Modal } from 'react-bootstrap'
-import React, { useState, useEffect } from 'react'
+import React, { useState } from 'react'
 import NewProjectForm from '../../Components/newProjectForm/NewProjectForm'
 import { useNavigate } from 'react-router-dom';
-import { updateActiveProject } from '../../Slices/projectSlice/activeProjectSlice';
-import { useDispatch, useSelector } from "react-redux";
 import '../../App.css'
 import './SelectProject.css'
-import { useFetch } from '../../shared/hooks/useFetch';
+import { useProjectsData } from './useProjectsData';
 
-
-
-
-const database = [
-];
 
 
 const SelectProject = () => {
-  const emailFromUser = useSelector((state) => state.user.user.Email);
   const navigate = useNavigate();
-  const dispatch = useDispatch();
-  const [viewModal, setViewModal] = useState(false);
-  const [projects, setProjects] = useState(database);
-  const { loading, error } = useFetch(`http://localhost:4000/projects/${emailFromUser}`, setProjects);
 
-  const submitNewProject = async (name, paymentPeriod) => {
-    const postFetch = await fetch("http://localhost:4000/projects", {
-      method: 'POST',
-      headers: {
-        "Content-type": "application/json",
-      },
-      body: JSON.stringify({
-        Nombre: name,
-        Periodo: paymentPeriod,
-        Email: emailFromUser,
-      }),
-    });
-    console.log(postFetch);
-  }
+  const [viewModal, setViewModal] = useState(false);
+  const { projects, handleProjectSelection, addNewEntry, loading, error } = useProjectsData();
+
+
 
   const handleCloseChild = () => {
     setViewModal(false);
   }
-
-
-  const addNewEntry = (newEntry) => {
-    setProjects([...projects, newEntry]);
-    submitNewProject(newEntry.Nombre, newEntry.TipoPeriodo);
-  }
-
-  const handleProjectSelection = (projectName) => {
-    console.log(`El seleccionado:${projectName}`);
-    dispatch(updateActiveProject(projectName));
-    navigate('/employees');
-
-  }
-
-
-
-
 
   return (
